- Fixed cross-compiling issues, added -m32 and -m64 flags to build scripts. - A number od cleanups to follow HICPPCM. - convert() template function for safe integer conversion. References #151
git-svn-id: https://devel.codership.com/svn/galera/trunk@1222 9148e567-b586-4176-b9b8-4e36cb4ec61b

+++ gcache/src/BufferHeader.hpp
@@ -19,7 +19,7 @@ namespace gcache
 
     struct BufferHeader
     {
-        int64_t  size;
+        ssize_t  size;
         int64_t  seqno;
         uint64_t flags;
     }__attribute__((__packed__));
